Move crosslang related exceptions to that module
